Fix popular song retrieval + generation
